src/detect_cat/requirements.txt
requests==2.25.1
Keras-Preprocessing==1.1.2
numpy==1.19.5